requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
APScheduler
attrs
blinker
cachetools
certifi
charset-normalizer
click
dataclasses-json
dnspython
faiss-cpu
filelock
filetype
Flask
frozenlist
fuzzywuzzy
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-generativeai
googleapis-common-protos
greenlet
grpcio
grpcio-status
gspread
gunicorn
h11
haversine
httpcore
httplib2
httpx
httpx-sse
idna
itsdangerous
Jinja2
jsonpatch
jsonpointer
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-text-splitters
langgraph
langgraph-checkpoint
langgraph-sdk
langsmith
MarkupSafe
marshmallow
msgpack
multidict
mypy-extensions
numpy
oauthlib
orjson
packaging
pandas
propcache
proto-plus
protobuf
pyasn1
pyasn1_modules
pydantic
pydantic-settings
pydantic_core
pymongo
pyparsing
python-dateutil
python-dotenv
pytz
PyYAML
RapidFuzz
requests
requests-oauthlib
requests-toolbelt
rsa
schedule
six
sniffio
SQLAlchemy
tenacity
tqdm
typing-inspect
typing_extensions
tzdata
tzlocal
uritemplate
urllib3
Werkzeug
yarl
spacy
fastapi
uvicorn
motor
pyngrok
langfuse
redis
google-cloud-storage
geopy
urllib3
instaloader
huey
httpx
